requirements.txt
fastapi
google-generativeai
uvicorn

nltk
numpy
pandas
pydantic